Level 9 (DoWithin Task and Find Task added, code quality improved)

+++ src/main/java/Lisa.java
@@ -1,24 +1,29 @@
-import exceptions.*;
-import classes.*;
-import enumerations.*;
+import exceptions.NoDescriptionException;
+import exceptions.NoTimeframeException;
+import exceptions.TooManySpacesException;
 import exceptions.Exception;
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
+import tasklist.TaskList;
+import storage.Storage;
+import ui.Ui;
+import enumerations.CommandEnum;
 
-/**
- * Lisa class. Stores tasks to a list and saves it to a txt file on device.
- */
+/** Lisa class. Stores tasks to a list and saves it to a txt file on device. */
 public class Lisa {
+    /** Handles any interaction or editing regarding the file. */
     private Storage storage;
+    /** Handles any interaction or editing regarding the list of tasks. */
     private TaskList tasks;
+    /** Handles any interaction regarding the terminal, e.g printing. */
     private Ui ui;
 
     /**
      * Constructor. Creates new Ui, Storage and Tasklist class for interaction with the terminal, txt file and
-     * the ArrayList of tasks
+     * the ArrayList of tasks.
      *
-     * @param filePath File path of the txt file (task_list.txt)
-     * @throws IOException Error for input-output
+     * @param filePath File path of the txt file (task_list.txt).
+     * @throws IOException Error for input-output.
      */
     public Lisa(String filePath) throws IOException {
         ui = new Ui();
@@ -27,27 +32,28 @@ public class Lisa {
     }
 
     /**
-     * Method to start running the programme. Accepts commands from user if matches with CommandEnum, otherwise prints
-     * error message with template for correction
+     * Runs the programme. Accepts commands from user if matches with CommandEnum, otherwise prints
+     * error message with template for correction.
      *
-     * @throws IOException Error for input-output
-     * @throws Exception Exception superclass. Has all the relevant self-created exceptions under it as subclasses
+     * @throws IOException Error for input-output.
+     * @throws Exception Exception superclass. Has all the relevant self-created exceptions under it as subclasses.
      */
     public void run() throws IOException, Exception {
         String input = "blank";
         while (!input.toLowerCase().equals("bye")) {
             input = ui.receive();
-            String[] temp = input.split(" ", 2);
+            String[] splitCmd = input.split(" ", 2);
             try {
-                CommandEnum cmd = CommandEnum.valueOf(temp[0].toUpperCase());
+                CommandEnum cmd = CommandEnum.valueOf(splitCmd[0].toUpperCase());
                 switch (cmd) {
                     case TODO:
                     case DEADLINE:
-                    case EVENT: {
+                    case EVENT:
+                    case WITHIN: {
                         try {
                             tasks.add(input);
-                            storage.insert(tasks.getStore().get(tasks.getStore().size()-1));
-                            ui.printAddMessage(tasks.getStore().get(tasks.getStore().size()-1));
+                            storage.insert(tasks.getStore().get(tasks.getStore().size() - 1));
+                            ui.printAddMessage(tasks.getStore().get(tasks.getStore().size() - 1));
                         } catch (NoDescriptionException e) {
                             ui.showNoDescriptionError();
                         } catch (NoTimeframeException e) {
@@ -65,11 +71,21 @@ public class Lisa {
                         ui.list(tasks);
                         break;
                     }
+                    case FIND: {
+                        try {
+                            ui.find(input, tasks);
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
+                        }
+                        break;
+                    }
                     case DAY: {
                         try {
                             ui.listByDay(input, tasks);
                         } catch (DateTimeParseException e) {
                             ui.showDateFormatError2();
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
                         }
                         break;
                     }
@@ -78,6 +94,8 @@ public class Lisa {
                             ui.listByMonth(input, tasks);
                         } catch (DateTimeParseException e) {
                             ui.showDateFormatError3();
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
                         }
                         break;
                     }
@@ -86,26 +104,40 @@ public class Lisa {
                             ui.listByYear(input, tasks);
                         } catch (DateTimeParseException e) {
                             ui.showDateFormatError4();
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
                         }
                         break;
                     }
                     case DELETE: {
                         try {
+                            if (tasks.getStore().size() == 0) {
+                                ui.printEmptyListMessage();
+                                break;
+                            }
                             tasks.delete(input);
                             storage.delete(input);
                             ui.printDeleteMessage();
-                        } catch (NumberFormatException | NullPointerException | IndexOutOfBoundsException | Exception e) {
+                        } catch (NumberFormatException | NullPointerException | IndexOutOfBoundsException e) {
                             ui.showInvalidNumberError();
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
                         }
                         break;
                     }
                     case DONE: {
                         try {
-                            tasks.done(input);
+                            tasks.setDone(input);
                             ui.printDoneMessage();
                         } catch (NumberFormatException | NullPointerException | IndexOutOfBoundsException e) {
                             ui.showInvalidNumberError();
+                        } catch (TooManySpacesException e) {
+                            ui.showTooManySpacesError();
                         }
+                        break;
+                    }
+                    case HELP: {
+                        ui.printCmdList();
                         break;
                     }
                     case BYE: {
@@ -123,8 +155,8 @@ public class Lisa {
      * Main function
      *
      * @param args
-     * @throws Exception All self-created errors
-     * @throws IOException Error for input-output
+     * @throws Exception All self-created errors. Refer to Exception class for specifics.
+     * @throws IOException Error for input-output.
      */
     public static void main(String[] args) throws Exception, IOException {
         new Lisa("data/task_list.txt").run();
